rename ResultsHelper package
