+++ External/Plugins/ResultsPanel/ResultsPanel.csproj
@@ -71,7 +71,7 @@
     <Compile Include="ListViewEx.cs">
       <SubType>Component</SubType>
     </Compile>
-    <Compile Include="Managers\ResultsPanelHelper.cs" />
+    <Compile Include="Helpers\ResultsPanelHelper.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="PluginMain.cs" />
     <Compile Include="PluginUI.cs">
